fixed the Decimal from the class Contract to BigDecial

+++ src/main/java/com/example/FinancialSystem/core/domain/classes/Contract.java
@@ -31,5 +31,5 @@ public class Contract {
 
     private BigDecimal monthlySetRate;
 
-    private Double remainingAmount;
+    private BigDecimal remainingAmount;
 }
